make test cleaner
