+++ engine/crates/integration-tests/tests/federation/basic/mod.rs
@@ -4,6 +4,7 @@
 //! that our engine supports all the things a normal GraphQL server should.
 
 // mod caching;
+mod empty_config;
 mod errors;
 mod fragments;
 mod headers;
@@ -15,50 +16,7 @@ mod variables;
 
 use engine_v2::Engine;
 use graphql_mocks::{FakeGithubSchema, MockGraphQlServer};
-use integration_tests::{
-    federation::{GatewayV2Ext, TestFederationGateway},
-    runtime,
-};
-use parser_sdl::federation::FederatedGraphConfig;
-use runtime::trusted_documents_client;
-use std::sync::Arc;
-
-#[test]
-fn works_with_empty_config() {
-    let federated_graph =
-        graphql_federated_graph::FederatedGraph::V3(graphql_federated_graph::FederatedGraphV3::default());
-
-    let cache = runtime_local::InMemoryCache::runtime(runtime::cache::GlobalCacheConfig {
-        enabled: true,
-        ..Default::default()
-    });
-
-    let federated_graph_config = FederatedGraphConfig::default();
-
-    let config = engine_config_builder::build_config(&federated_graph_config, federated_graph);
-    let gateway = TestFederationGateway {
-        gateway: Arc::new(engine_v2::Engine::new(
-            engine_v2::Schema::try_from(config.into_latest()).unwrap(),
-            engine_v2::EngineEnv {
-                fetcher: runtime_local::NativeFetcher::runtime_fetcher(),
-                cache: cache.clone(),
-                trusted_documents: trusted_documents_client::Client::new(
-                    runtime_noop::trusted_documents::NoopTrustedDocuments,
-                ),
-                kv: runtime_local::InMemoryKvStore::runtime(),
-                meter: grafbase_tracing::metrics::meter_from_global_provider(),
-            },
-        )),
-    };
-
-    let request: engine::Request = serde_json::from_value(serde_json::json!({"query": "{ __typename }"})).unwrap();
-
-    runtime().block_on(
-        gateway
-            .gateway
-            .execute(Default::default(), engine::BatchRequest::Single(request)),
-    );
-}
+use integration_tests::{federation::GatewayV2Ext, runtime};
 
 #[test]
 fn single_field_from_single_server() {
